Fixed invalid cast warnings under Linux.
git-svn-id: file:///home/michael/svn/yuni/trunk@1250 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/directory/copy.cpp
@@ -39,9 +39,9 @@ namespace Directory
 	{
 		// normalize paths
 		String fsrc(src, srclen);
-		Yuni::Core::IO::Normalize(fsrc, src, srclen);
+		Yuni::Core::IO::Normalize(fsrc, src, (unsigned int)srclen);
 		String fdst(dst, dstlen);
-		Yuni::Core::IO::Normalize(fdst, dst, dstlen);
+		Yuni::Core::IO::Normalize(fdst, dst, (unsigned int)dstlen);
 
 		// The list of files to copy
 		List list;
